This commit fixes a bug in commit 1542fcd where I forgot to update index.html with the new name of the file.
check_questions_hack.js --> hack_add_questions.js inside of index.html is the change which fixes the bug "hack3_main is not defined" and allows the player to actually get stuck in the forever loop.

With this fix, Hack3 is now functioning normally and can be tested by adding 5 questions to the forever loop.