This commit changes filenames associated with hack2.
The convention for any given hack number X is:
name_of_hack_hackX.js
hackX_story.js
hackX_check.js

We are in the process of updating hacks 1 and 3 to match this structure accordingly.

Also, WHENEVER WE CHANGE FILENAMES IN THIS REPO, we MUST also change the src in index.hmtl so the game knows where to find the file. So, this commit also changes the appropriate script tags in index.html with the new filenames.